Struct loopdev::LoopDevice

pub struct LoopDevice {
    // some fields omitted
}

Interface to a loop device ie /dev/loop0.

Methods

impl LoopDevice

fn open(dev: &str) -> Result<LoopDevice>

Opens a loop device.

fn attach(&self, backing_file: &str, offset: u64) -> Result<()>

Attach the loop device to a file starting at offset into the file.

Examples

Attach the device to the start of a file.

use loopdev::LoopDevice;
let ld = LoopDevice::open("/dev/loop4").unwrap();
ld.attach("test.img", 0).unwrap();

fn get_path(&self) -> Option<PathBuf>

Get the path of the loop device.

fn detach(&self) -> Result<()>

Detach a loop device from its backing file.

Examples

use loopdev::LoopDevice;
let ld = LoopDevice::open("/dev/loop6").unwrap();
ld.detach().unwrap();