Struct loopdev::LoopDevice

pub struct LoopDevice { /* fields omitted */ }

Interface to a loop device ie /dev/loop0.

Methods

impl LoopDevice

pub fn open<P: AsRef<Path>>(dev: P) -> Result<LoopDevice>

Opens a loop device.

pub fn attach<P: AsRef<Path>>(&self, backing_file: P, offset: u64) -> Result<()>

Attach the loop device to a file starting at offset into the file.

Examples

Attach the device to the start of a file.

use loopdev::LoopDevice;
let ld = LoopDevice::open("/dev/loop4").unwrap();
ld.attach("test.img", 0).unwrap();

pub fn get_path(&self) -> Option<PathBuf>

Get the path of the loop device.

pub fn detach(&self) -> Result<()>

Detach a loop device from its backing file.

Examples

use loopdev::LoopDevice;
let ld = LoopDevice::open("/dev/loop6").unwrap();
ld.detach().unwrap();